Add deprecated headers support to syncqt
When renaming classes, or when moving classes from one module to
another, it's useful to have a simple way of supporting the old
API/location for some time. To this end, syncqt shall now recognize
a "deprecatedheaders" section in sync.profile. It looks like this:

%deprecatedheaders = (
    "QtDeclarative" => {
        "qquickcanvas.h" => "QtQuick2/qquickcanvas.h",
        "qquickitem.h" => "QtQuick2/qquickitem.h",
        "QQuickCanvas" => "QtQuick2/QQuickCanvas",
        "QQuickItem" => "QtQuick2/QQuickItem",
    }
);

In the above example, syncqt would generate a header called
qquickcanvas.h for the QtDeclarative module; when included, this
header will issue a warning and include <QtQuick2/qquickcanvas.h>.
And so on, for the other entries.

Deprecated headers are installed along with the module's normal
headers.

Change-Id: Ie2518b42275c2b2ff44216f07d376ccf5be6dc45
Reviewed-by: Marius Storm-Olsen <marius.storm-olsen@nokia.com>